allow join, writer deny by role

+++ src/main/java/net/grayfield/spb/hobbylog/domain/walk/struct/Walk.java
@@ -11,6 +11,6 @@ import org.springframework.data.mongodb.core.mapping.Document;
 public class Walk extends BaseSchema {
     private String content;
     private int steps;
-    private int distance;
+    private float distance;
     private int duration;
 }
